backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.11.7
pydantic-settings==2.1.0

# LangChain and RAG dependencies
langchain==0.1.0
langchain-community==0.0.10
google-genai==0.2.2
chromadb==0.4.22
sentence-transformers==2.2.2

# Document processing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
requests==2.32.3
colorama==0.4.6
